Hyprland: Add option to declare path for hyprctl
Home Manager gives an error if package and portal package are set to null but you have plugins declared due to the new exec-once change. This adds an option that is null by default to decalre a path to hyprctl, when the option is null it uses the normal cfg.finalPackage path, but when declared it uses that declared path.